% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSmake_tcplotdata.R
\name{BSmake_tcplotdata}
\alias{BSmake_tcplotdata}
\title{time course plots for eye tracking computes avg proportion and BOOTSTRAPPED by subject 95\% CIs (for Dale).}
\usage{
BSmake_tcplotdata(df, AOIs, subj, n = 100, ..., bin = CURRENT_BIN,
  bin_interval = 20, level = 0.95)
}
\arguments{
\item{df}{dataframe}

\item{AOIs}{columnnames of AOIs (for eyelink trackers, the RIGHT_1_P, RIGHT_2_P,... columns)}

\item{subj}{participant column}

\item{n}{bootstrap samples. so.. R really..}

\item{...}{variable names of experimental conditions}

\item{bin}{BIN variable (for eyelink, this is the CURRENT_BIN column). Defaults to "CURRENT_BIN".}

\item{bin_interval}{the time of each bin}

\item{level}{confidence level, .95 by default}
}
\description{
time course plots for eye tracking computes avg proportion and BOOTSTRAPPED by subject 95\% CIs (for Dale).
}
\examples{
plottingdata <- BSmake_tcplotdata(df,AOIs=c(refprop,disprop),subj=Participant,fluency,gesture,n=1000)
tcplot(plottingdata,0,2000,lty=fluency)+facet_wrap(~gesture)
}
