% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proptable.R
\name{create_proptable}
\alias{create_proptable}
\title{proportion tibbles}
\usage{
create_proptable(freq, var1, data)
}
\arguments{
\item{freq}{categorical response variable (e.g. frequency data)}

\item{var1}{grouping variable accross which frequency variable gets proportioned of a total n in that group.}
}
\description{
proportion tibbles
}
\examples{
proptables<-create_proptable("stroke","age_group",data))
ggplot(proptables, aes(x=stroke, y=prop,fill=stroke)) + 
ylim(0,1)+
geom_bar(position=position_dodge(), stat="identity")+facet_wrap(~(age_group))+
geom_errorbar(aes(ymin=intlow, ymax=inthigh),width=.2,position=position_dodge(.9))

or, just for response of 1 (i.e. just plot bar for occurrences of stroke, not no-occurrence of stroke)
ggplot(proptables[proptable$stroke==1,], aes(x=age_group, y=prop)) + 
ylim(0,1)+
geom_bar(position=position_dodge(), stat="identity")+
geom_errorbar(aes(ymin=intlow, ymax=inthigh),width=.2,position=position_dodge(.9))
}
