% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquaredfunction.R
\name{.rsquared.glmm}
\alias{.rsquared.glmm}
\title{Marginal and conditional r-squared for glmm given fixed and random variances}
\usage{
.rsquared.glmm(varF, varRand, varResid = NULL, varDisp = NULL, family, link,
  mdl.aic, mdl.class, null.fixef = NULL)
}
\arguments{
\item{varF}{Variance of fixed effects}

\item{varRand}{Variance of random effects}

\item{varResid}{Residual variance. Only necessary for "gaussian" family}

\item{family}{family of the glmm (currently works with gaussian, binomial and poisson)}

\item{link}{model link function. Working links are: gaussian: "identity" (default);
binomial: "logit" (default), "probit"; poisson: "log" (default), "sqrt"}

\item{mdl.aic}{The model's AIC}

\item{mdl.class}{The name of the model's class}

\item{null.fixef}{Numeric vector containing the fixed effects of the null model.
Only necessary for "poisson" family}
}
\value{
A data frame with "Class", "Family", "Marginal", "Conditional", and "AIC" columns
}
\description{
This function is based on Nakagawa and Schielzeth (2013). It returns the marginal
and conditional r-squared, as well as the AIC for each glmm.
Users should call the higher-level generic "r.squared", or implement a method for the
corresponding class to get varF, varRand and the family from the specific object
}
