% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_columns}
\alias{find_columns}
\title{Finds named columns or fields within workspace}
\usage{
find_columns(name)
}
\arguments{
\item{name}{a character vector of names to find}
}
\value{
a list of objects giving both the object name and its value
}
\description{
Finds columns in data frames or entries in lists by name.  Often
you want to find a specific named column but can't remember which
data frame it was held in.
}

