% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCaterpillar.R
\name{ggCaterpillar}
\alias{ggCaterpillar}
\title{caterpillar plot for mixed effects models
borrowed (stolen) from http://stackoverflow.com/questions/13847936/in-r-plotting-random-effects-from-lmer-lme4-package-using-qqmath-or-dotplot}
\usage{
ggCaterpillar(re, QQ = TRUE, likeDotplot = TRUE)
}
\description{
caterpillar plot for mixed effects models
borrowed (stolen) from http://stackoverflow.com/questions/13847936/in-r-plotting-random-effects-from-lmer-lme4-package-using-qqmath-or-dotplot
}
\examples{
ggCaterpillar(ranef(model, postVar=TRUE))
}
