% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{grid arrange (from gridExtra) with common legend}
\usage{
grid_arrange_shared_legend(
  ...,
  nrow = 1,
  ncol = length(list(...)),
  position = c("bottom", "right")
)
}
\arguments{
\item{...}{list of ggplot objects}

\item{nrow}{number of plot rows}

\item{ncol}{number of plot columns (defaults to nr of plots)}

\item{position}{position of legend.}
}
\description{
grid arrange (from gridExtra) with common legend
}
\examples{
grid_arrange_shared_legend(p1,p2,p3,p4, nrow=2,ncol=2)
}
