% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{jkcolours}
\alias{jkcolours}
\title{Colours}
\format{\preformatted{List of 6
 $ impblue      : chr "#156A9B"
 $ uesgreen     : chr "#22B14C"
 $ bpyellow     : chr "#FCE911"
 $ bpgreenyellow: chr "#ACC312"
 $ implightblue : chr "#8BAECC"
 $ impbrown     : chr "#9C4716"
}}
\usage{
jkcolours
}
\value{
a list with the following defined colours:
\itemize{
\item \code{impblue} Imperial College dark blue
\item \code{implightblue} Imperial College light blue
\item \code{impbrown} complement of \code{impblue}
\item \code{uesgreen} green of Urban Energy Systems project logo
\item \code{bpyellow} yellow from BP logo
\item \code{bpgreenyellow} green-yellow from BP logo
}
}
\description{
Frequently used colour definitions.
}
\keyword{datasets}

