% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{jkeirstead}
\alias{jkeirstead}
\alias{jkeirstead-package}
\title{jkeirstead}
\description{
This package provides convenient access to functions and variables
that I commonly use for R analysis.  By putting these functions
into their own package, it becomes easier to share code rather than
passing around customized .Rprofile files and I can still
automatically load everything but with a simpler
\code{library(jkeirstead)} call.
}

