% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.beta.lmer.R
\name{lm.beta.lmer}
\alias{lm.beta.lmer}
\title{equivalent of lm.beta for mixed models.}
\usage{
lm.beta.lmer(mod)
}
\arguments{
\item{mod}{linear mixed model object}
}
\description{
equivalent of lm.beta for mixed models.
}
\examples{
lm.beta.lmer(fit1)
}
