% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lm_to_latex}
\alias{lm_to_latex}
\title{Converts a model formula to LaTeX code}
\usage{
lm_to_latex(model, mathmode = FALSE)
}
\arguments{
\item{model}{a model object}

\item{mathmode}{boolean indicating whether the formula will be
displayed in math mode.  Default = false.}
}
\value{
a character string formatted for LaTeX (with knitr,
i.e. multiple escapes).  Users can add mathmode or texttt as they
wish from knitr.
}
\description{
Converts the formula of a model object to LaTeX code for use in a
knitr report or similar.
}

