% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_MC_sample}
\alias{make_MC_sample}
\title{Makes a Monte Carlo sample using Sobol' sequences}
\usage{
make_MC_sample(n, vals)
}
\arguments{
\item{n}{number of samples to draw}

\item{vals}{a list describing distributions.  Each list entry
should consist of the following named fields:

\itemize{

\item \code{var} a variable name

\item \code{dist} the distribution name, e.g. 'unif' for Uniform
(see R's distribution names)

\item \code{params} a list of distribution parameters.  The list
names will vary by distribution.

}}
}
\value{
a data frame with an index column "n" and each distribution
in a named column
}
\description{
Generates a Monte Carlo sample using Sobol' low-discrepancy
quasi-random sequences. Random sampling with R's standard methods
is inefficient for Monte Carlo analysis as the sampled values do
not cover the parameter space evenly.  This function allows users
to create parameter samples using quasi-random low-discrepancy
Sobol' sequences to get around this problem.
}
\examples{
{
## Define the distributions
vals <- list(list(var="Uniform",
              dist="unif",
              params=list(min=0,max=1)),
         list(var="Normal",
              dist="norm",
              params=list(mean=0,sd=1)),
         list(var="Weibull",
              dist="weibull",
              params=list(shape=2,scale=1)))

## Generate 100 samples
samp <- make_MC_sample(100, vals)
}
}
\seealso{
\url{https://gist.github.com/jkeirstead/1730440}
}

