% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tcplot_data.R
\name{make_tcplot_data}
\alias{make_tcplot_data}
\title{time course plots for eye tracking}
\usage{
make_tcplot_data(objects, predictor, bin = "CURRENT_BIN", data)
}
\arguments{
\item{objects}{vector of proportions of fixations to visual world AOIs (for eyelink trackers, the RIGHT_1_P, RIGHT_2_P,... columns)}

\item{predictor}{variable name (string) of experimental conditions.}

\item{bin}{BIN variable (for eyelink, this is the CURRENT_BIN column). Defaults to "CURRENT_BIN".}

\item{data}{dataset}
}
\description{
time course plots for eye tracking
}
\examples{
plottingdata <- make_tcplot_data(objects=c("easy_fix","diff_fix","vid_fix"),predictor="g_condition",data=e6data)
}
