% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tcplotdata.R
\name{make_tcplotdata}
\alias{make_tcplotdata}
\title{time course plots for eye tracking computes avg proportion and by subject standard errors.}
\usage{
make_tcplotdata(df, AOIs, subj, ..., bin = CURRENT_BIN,
  bin_interval = 20)
}
\arguments{
\item{df}{dataframe}

\item{AOIs}{column names of AOIs (for eyelink trackers, the RIGHT_1_P, RIGHT_2_P,... columns)}

\item{subj}{participant column}

\item{...}{variable names of experimental conditions}

\item{bin}{BIN variable (for eyelink, this is the CURRENT_BIN column). Defaults to "CURRENT_BIN".}

\item{data}{dataset}
}
\description{
time course plots for eye tracking computes avg proportion and by subject standard errors.
}
\examples{
plottingdata <- make_tcplotdata(df,AOIs=c(refprop,disprop),subj=Participant,fluency, gesture)
}
