% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{rounding to the nearest}
\usage{
mround(x, base)
}
\arguments{
\item{x}{numeric value to round}

\item{base}{to the nearest}
}
\description{
rounding to the nearest
}
\examples{
mround(1002,20)
}
