% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eyes.R
\name{plot_eyes}
\alias{plot_eyes}
\title{time course plots for eye tracking}
\usage{
plot_eyes(objects, predictor, xmin_time = 0, xmax_time = 2000,
  bin = "CURRENT_BIN", bin_interval = 20, data, facet_var = "none")
}
\arguments{
\item{objects}{vector of proportions of fixations to visual world AOIs (for eyelink trackers, the RIGHT_1_P, RIGHT_2_P,... columns)}

\item{predictor}{variable name (string) of experimental conditions.}

\item{xmin_time}{start time of plot. Defaults to 0.}

\item{xmax_time}{end time of plot. Defaults to 2000.}

\item{bin}{BIN variable (for eyelink, this is the CURRENT_BIN column). Defaults to "CURRENT_BIN".}

\item{bin_interval}{duration of each bin in ms. Defaults to 20.}

\item{data}{dataset}

\item{facet_var}{The variable by which to facet plots. Defaults to "none".}
}
\description{
time course plots for eye tracking
}
\examples{
plot_eyes(objects=c("easy_fix","diff_fix","vid_fix"),predictor="g_condition",xmin_time=0,xmax_time=4000,data=e6data,facet_var="predictor")
}
