% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prose_vector}
\alias{prose_vector}
\title{Converts a character vector to readable prose}
\usage{
prose_vector(vals, oxford = FALSE, and = "and")
}
\arguments{
\item{vals}{a character vector}

\item{oxford}{a boolean indicating whether Oxford comma should be
used.  If true a comma is placed before the \code{and} word.}

\item{and}{word before penultimate entry.  Default = "and"}
}
\value{
a character of length one
}
\description{
Converts a character vector to a prose list.  Often when working
with reproducible reports, a user will want to print a list of
character values but needs to respect English grammar: for example,
\code{"dog", "cat", "mouse"} should be rendered as "dog, cat and
mouse".
}
\seealso{
https://gist.github.com/jkeirstead/48bf5845b235dc7a8d5a
}

