% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsample.R
\name{qsample}
\alias{qsample}
\title{quick sample of .. see https://www.tjmahr.com/sample-n-groups/}
\usage{
qsample(data, size, ...)
}
\arguments{
\item{size}{size of sample, passed to sample() function.}

\item{...}{grouping variables. samples from all available combinations}

\item{df}{dataframe/tibble}
}
\description{
quick sample of .. see https://www.tjmahr.com/sample-n-groups/
}
\examples{
qsample(df,4,sub_id) \%>\% ggplot(......
qsample(df,4,sub_id, trial_id) \%>\% ggplot(......
}
