% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_drive.R
\name{read_from_drive}
\alias{read_from_drive}
\title{read's csv from drive file.}
\usage{
read_from_drive(id_pattern = "null")
}
\arguments{
\item{id_pattern}{regex pattern to partial match filename in drive.}
}
\description{
read's csv from drive file.
}
\examples{
read_from_drive("E6")
}
