% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSums_na.R
\name{rowSums_na}
\alias{rowSums_na}
\title{rowSums with na.rm = TRUE except for when all cols = NAs
equivalent to sum functionality in SAS (I think)}
\usage{
rowSums_na(x, dims = 1L)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric, complex, integer or logical values, or a numeric data frame. For .colSums() etc, a numeric, integer or logical matrix (or vector of length m * n).}

\item{dims}{integer: Which dimensions are regarded as ‘rows’ or ‘columns’ to sum over. For row*, the sum or mean is over dimensions dims+1, ...; for col* it is over dimensions 1:dims.}
}
\description{
rowSums with na.rm = TRUE except for when all cols = NAs
equivalent to sum functionality in SAS (I think)
}
\examples{
x <- cbind(x1 = c(NA,2:5), x2 = c(NA,7:9,NA))
rowSums_na(x)
## as opposed to 
rowSums(x)

x <- array(2, rep(3, 3))
x[1,,]<-NA
x[3,3,3]<-NA
rowSums_na(x)
## as opposed to 
rowSums(x)
}
