% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.stat.R
\name{t.stat}
\alias{t.stat}
\title{extract t values for fixed effects of lmer object}
\usage{
\method{t}{stat}(x)
}
\arguments{
\item{x}{a linear mixed effects model object.}
}
\description{
extract t values for fixed effects of lmer object
}
\examples{
t.stat(fittedmodel1)
}
