% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplot.R
\name{tcplot}
\alias{tcplot}
\title{time course plots for eye tracking
defaults to colour by AOI.}
\usage{
tcplot(df, xmin = 0, xmax = 2000, col = AOI, lty = NULL,
  x = time, y = mean_prop, ymin = low, ymax = up)
}
\arguments{
\item{df}{a dataframe in long format from the make_tcplotdata function}

\item{xmin, xmax}{min and max time boundaries. (ROI/IP)}

\item{col}{= line colours variable name}

\item{lty}{= line type variable name}
}
\description{
time course plots for eye tracking
defaults to colour by AOI.
}
\examples{
tcplot(plotting_data, -1000, 1000)
tcplot(plotting_data, -1000, 1000,col=fluency, lty=AOI)
}
