% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplot_nolines.R
\name{tcplot_nolines}
\alias{tcplot_nolines}
\title{time course plots for eye tracking}
\usage{
tcplot_nolines(df, xmin = 0, xmax = 2000, lcol = "AOI")
}
\arguments{
\item{df}{a dataframe in long format from the make_tcplot_data function}

\item{xmin, xmax}{min and max time boundaries. (ROI)}

\item{lcol}{= line colours variable, string.}

\item{ltype}{= line type variable, string.}
}
\description{
time course plots for eye tracking
}
\examples{
tcplot_nolines(plotting_data, -1000, 4000)
tcplot_nolines(plotting_data, -1000, 4000)+facet_wrap(~condition)
}
