% Generated by roxygen2 (4.0.1): do not edit by hand
\name{theme_schematic}
\alias{theme_schematic}
\title{A simple theme for schematics}
\usage{
theme_schematic()
}
\description{
A simple ggplot2 theme for schematics.  Sometimes a person wants to
plot a general curve but without paying particular attention to the
axis values.  This is a very stripped back theme that removes axis
scale data and tick marks.
}

