% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoepsy_rmd_funcs.R
\name{uoepsy_rmd_funcs}
\alias{uoepsy_rmd_funcs}
\title{get a quick list of all functions used in Rmd. Will exclude echo=FALSE chunks on premise that students aren't presented with these, so we often put more complicated plotting stuff in there.}
\usage{
uoepsy_rmd_funcs(filepath, excludechunks = "echo=FALSE")
}
\arguments{
\item{filepath}{rmd document for lab/lecture}
}
\description{
get a quick list of all functions used in Rmd. Will exclude echo=FALSE chunks on premise that students aren't presented with these, so we often put more complicated plotting stuff in there.
}
\examples{
uoepsy_rmd_funcs("01-slr.Rmd")
}
