% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{negation of %in% base function}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{the values to be matched}

\item{y}{the values to be matched against}
}
\description{
negation of %in% base function
}
\examples{
1:4 \%notin\% 2:5 returns TRUE, FALSE, FALSE, FALSE
}
