% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tcplot_data.R
\name{make_tcplot_data}
\alias{make_tcplot_data}
\title{time course plots for eye tracking}
\usage{
make_tcplot_data(df, AOIs, predictor, bin = "CURRENT_BIN",
  bin_interval = 20)
}
\arguments{
\item{AOIs}{vector of string names of AOIs (for eyelink trackers, the RIGHT_1_P, RIGHT_2_P,... columns)}

\item{predictor}{variable name (string) of experimental condition.}

\item{bin}{BIN variable (for eyelink, this is the CURRENT_BIN column). Defaults to "CURRENT_BIN".}

\item{data}{dataset}
}
\description{
time course plots for eye tracking
}
\examples{
plottingdata <- make_tcplot_data(df=e6data, objects=c("easy_fix","diff_fix","vid_fix"),predictor=gesture)
}
