% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplot.R
\name{tcplot}
\alias{tcplot}
\title{time course plots for eye tracking}
\usage{
tcplot(df, xmin = 0, xmax = 2000, lcol = condition, ltype = object)
}
\arguments{
\item{df}{a dataframe in long format from the make_tcplot_data function}

\item{xmin, xmax}{min and max time boundaries. (ROI)}

\item{lcol}{= line colours variable.}

\item{ltype}{= line type variable.}
}
\description{
time course plots for eye tracking
}
\examples{
tcplot(plotting_data, -1000, 4000)
tcplot(plotting_data, -1000, 4000)+facet_wrap(~condition)
}
